Séparation des logiques de route et ajout de SWAGGER
